# Implement faster NNLS solver based on the Goldfarb/Idnani dual algorithm