--- a/docsrc/source/installation.rst
+++ b/docsrc/source/installation.rst
@@ -45,7 +45,7 @@ DeerLab installs the following packages:
 * `joblib <https://joblib.readthedocs.io/en/latest/>`_ - Library lightweight pipelining and parallelization.
 * `tqdm <https://github.com/tqdm/tqdm>`_ - A lightweight package for smart progress meters.
 * `dill <https://github.com/uqfoundation/dill>`_ - An extension of Python's pickle module for serializing and de-serializing python objects.
-
+* `quadprog <https://pypi.org/project/quadprog/>`_ - A quadratic programming solver
 Installing from Anaconda
 *************************