Implements a new NNLS solver function `qpnnls` that uses the Goldfarb/Idnani dual algorithm implemented in the `quadprog` package. The resulting NNLS solver is more efficient than the current implementation with `cvxopt` (see below for a benchmark example) without sacrificing any (noticeable) accuracy (I tested and benchmarked this to make sure). Since NNLS problem-solving is the bottleneck of most applications of DeerLab, it results in a significant improvement in speed when analyzing semi-parametric models. 

The following benchmarks the performance of DeerLab's `cvxnnls` and `qpnnls` functions when solving the NNLS problem of the dipolar model. The analysis is repeated with different `rmax` values of the distance vector with a fixed resolution of 0.05nm (an approximate lower bound of physically-meaningful resolution attainable), effectively increasing the number of linear parameters to estimate during the NNLS problem.
 
```python 
import numpy as np
import deerlab as dl 
import matplotlib.pyplot as plt 
from tqdm import tqdm 
from deerlab.solvers import cvxnnls, qpnnls
from timeit import timeit 

np.random.seed(42)
Nrs = np.arange(50,200,2)
cvx_times,qp_times,cvx_rmsd,qp_rmsd = [],[],[],[]
for nr in tqdm(Nrs):

    # Dipolar EPR NNLS problem
    dr = 0.05
    t = np.arange(-1,5,0.008)
    r = np.linspace(1.5,2+dr*nr,nr)
    A = dl.dipolarkernel(t,r)
    alpha=0.005
    L = dl.regoperator(r,2)
    xtruth = dl.dd_gauss(r,3,0.2)
    # Forward simulation
    y = A@xtruth + dl.whitegaussnoise(t,0.05)

    # Prepare NNLS problem
    AtAreg = A.T@A + alpha**2*L.T@L
    Aty = A.T@y
    N = len(r)
    
    # Solve NNLS problem
    fcn = lambda : qpnnls(AtAreg, Aty)
    duration = timeit(fcn, number=100)
    qp_times.append(duration)
    qp_rmsd.append(1/nr*np.linalg.norm(xtruth-fcn()))

    fcn = lambda: cvxnnls(AtAreg,Aty)
    duration = timeit(fcn, number=100)
    cvx_times.append(duration)
    cvx_rmsd.append(1/nr*np.linalg.norm(xtruth-fcn()))


rmax = 2+dr*Nrs
#plt.subplot(121)
plt.plot(rmax,cvx_times,label='CVX-NNLS')
plt.plot(rmax,qp_times,label='QP-NNLS')
plt.yscale('log')
plt.ylabel('Run time [s]')
plt.xlabel('rmax [nm]')
plt.legend(loc='best')
```

![Figure_1](https://user-images.githubusercontent.com/48292540/197790430-e9de233d-4c1b-4c7f-93ba-84a0035dbef7.png)

The results show that the `qpnnls` is significantly faster than `cvxopt` for most physically-meaningful cases here. If the distance range is significantly constrained (i.e. the number of linear parameter is small), the algorithm is orders of magnitude faster. 

Another indicator is the duration of the `pytest` run over the full test suite. It takes `pytest` approximately 232s using `cvxopt` and 125s using `qpnnls` showing a general doubling of the execution speed. For the `snlls` test we also observe the increase in performance:
| Test            | `cvxnnls` | `qpnnls` |
|-----------------|---------|--------|
| `test_snlls_rlls` | 32.49s   | 13.61s  |
| `test_snlls_nlls` | 4.20s   | 4.05s   |
| `test_snlls`      | 45.43s   | 16.23s  |

Due to the overall good performance of the `qpnnls` solver, the PR will set it as the default NNLS solver engine, whereas `cvxnnls` will remain accessible through the `nnlsSolver` option of `snlls`.